requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-graphql-jwt==0.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
idna==3.10
promise==2.3
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
urllib3==2.5.0
